+++ test/query/res_builders_test.dart
@@ -29,12 +29,12 @@ void main() {
       )
       ..list('list1', attrs, '201', 'test-page-id', (rb) {
         rb
-          ..item('xid-3', 'actor3', (rb) {
+          ..item(3.0, 'actor3', (rb) {
             rb
               ..val('view100', 'value3100', '3100')
               ..val('view110', 'value3110', '3110');
           })
-          ..item('xid-4', 'actor4', (rb) {
+          ..item(4.0, 'actor4', (rb) {
             rb
               ..val('view100', 'value4100', '4100')
               ..val('view110', 'value4110', '4110');
@@ -92,12 +92,12 @@ void main() {
     var list1 = res.views['list1'] as ListQueryResult;
     expect(list1, isNotNull);
     expect(list1, isA<ListQueryResult>());
-    // Check ListItems with keys and values
+    // Check ListItems with positions and refIds
     expect(list1.value.length, 2);
-    expect(list1.value.elementAt(0).key, 'xid-3');
-    expect(list1.value.elementAt(0).value, 'actor3');
-    expect(list1.value.elementAt(1).key, 'xid-4');
-    expect(list1.value.elementAt(1).value, 'actor4');
+    expect(list1.value.elementAt(0).position, 3.0);
+    expect(list1.value.elementAt(0).refId, 'actor3');
+    expect(list1.value.elementAt(1).position, 4.0);
+    expect(list1.value.elementAt(1).refId, 'actor4');
     expect(list1.changeId, '201');
     expect(list1.attrs, {
       'actor3': {
@@ -176,12 +176,12 @@ void main() {
       )
       ..list('list1', attrs, '201', 'test-page-id', (rb) {
         rb
-          ..item('xid-3', 'actor3', (rb) {
+          ..item(3.0, 'actor3', (rb) {
             rb
               ..val('view100', 'value3100', '3100')
               ..val('view110', 'value3110', '3110');
           })
-          ..item('xid-4', 'actor4', (rb) {
+          ..item(4.0, 'actor4', (rb) {
             rb
               ..val('view100', 'value4100', '4100')
               ..val('view110', 'value4110', '4110');
@@ -209,8 +209,8 @@ void main() {
       'list1': {
         'type': 'list',
         'val': [
-          {'key': 'xid-3', 'value': 'actor3'},
-          {'key': 'xid-4', 'value': 'actor4'},
+          {'position': 3.0, 'refId': 'actor3'},
+          {'position': 4.0, 'refId': 'actor4'},
         ],
         'attrs': {
           'actor3': {